Start of filter by event type
I didn't see any way of filtering the query to the github api by event type. I decided to show the list of returned events and filter them in the javascript code. I only got as far as rendering a list of event types to filter by and ran out of time before actually creating the filter component and logic.